don't download arch manifest if it already exists
  Arch manifests should not ever change, so downloading fresh copies is
  a waste of time.

  Added test coverage for manifest download and failure states.

+++ raindance/package.py
@@ -37,11 +37,15 @@ class PackageArchive(object):
 
     def save_arch_manifest(self, software, version, verdir, arch):
         afname = '{}.json'.format(arch)
-        archurl = path(self.root_url) / software / version / afname
         archfile = verdir / afname
+        if archfile.exists():
+            return archfile, json.loads(archfile.text())
+
+        archurl = path(self.root_url) / software / version / afname
         res = self.http.get(str(archurl))
         if not res.ok:
             raise RuntimeError('Request for %s failed: %s',  archurl, res)
+
         archfile.write_text(res.text)
         return archfile, res.json()
